requirements.txt
numpy==1.25.2
pandas==2.2.2
scikit-learn==1.2.2
scikit-image==0.20.0
matplotlib==3.7.2
scipy==1.10.1
opencv-python==4.8.1.78
Pillow==10.0.1
kneed==0.8.1
mplcursors==0.6
umap-learn==0.5.4
mahotas==1.4.13
tqdm>=4.0
xgboost
